frontend: rename schema to frontend
