+++ services/frontend/next.config.js
