+++ services/frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "schema",
+  "name": "frontend",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
