+++ services/frontend/pages/builder.js
@@ -7,7 +7,7 @@ const floatStyle = {
     position: "fixed",
     right: "5%",
     bottom: "5%"
-  }
+}
 
 class field {
     constructor(id, name, type, optional, indents) {
@@ -62,8 +62,8 @@ export default class Builder extends React.Component {
     render() {
         return (
             <Layout>
-                <h1 style={{color: "white"}}>CRD Schema Builder</h1>
-                <p style={{color: "white"}}>Design your schema then generate your CustomResourceDefintion.</p>
+                <h1 style={{ color: "white" }}>CRD Schema Builder</h1>
+                <p style={{ color: "white" }}>Design your schema then generate your CustomResourceDefintion.</p>
                 {this.state.inputs.map(f =>
                     <FieldInput key={f.id} field={f} updater={this.updateField} deleter={this.deleteField} adder={this.addField}></FieldInput>
                 )}
